modity code structure
