+++ src/llmtuner/api/__init__.py
@@ -1,0 +1,1 @@
+from llmtuner.api.app import create_app
