+++ src/llmtuner/extras/constants.py
@@ -1,0 +1,7 @@
+IGNORE_INDEX = -100
+
+VALUE_HEAD_FILE_NAME = "value_head.bin"
+
+FINETUNING_ARGS_NAME = "finetuning_args.json"
+
+LAYERNORM_NAMES = ["norm", "ln_f", "ln_attn", "ln_mlp"] # for LLaMA, BLOOM and Falcon settings
