+++ src/llmtuner/tuner/pt/__init__.py
@@ -1,0 +1,1 @@
+from llmtuner.tuner.pt.workflow import run_pt
