+++ src/llmtuner/tuner/rm/__init__.py
@@ -1,0 +1,1 @@
+from llmtuner.tuner.rm.workflow import run_rm
