+++ src/llmtuner/tuner/sft/__init__.py
@@ -1,0 +1,1 @@
+from llmtuner.tuner.sft.workflow import run_sft
